+++ src/validate-package-string-configuration.ts
@@ -2,6 +2,7 @@ import { ValidationErrorCodes } from "./error-codes";
 import { error, success } from "./chalk-formatting";
 import { isMetaPackage } from "./is-meta-package";
 import { ValidateMetaPackageFunctionDefinition } from "./validate-meta-package-function-definition";
+import { buildFullPackageDescription } from "./build-full-package-description";
 
 /** Validates the string content of a `meta-package.json` file */
 export const validatePackageStringConfiguration = async (configurationData : string) => {
@@ -14,7 +15,9 @@ export const validatePackageStringConfiguration = async (configurationData : str
   }
 
   isMetaPackage(objectResult);
-  await new ValidateMetaPackageFunctionDefinition(objectResult)
+  const builtMetaPackage = await buildFullPackageDescription(objectResult);
+
+  await new ValidateMetaPackageFunctionDefinition(builtMetaPackage)
     .execute();
 
   console.log(success("Final Package file passed validation."));